requirements.txt
mysql-connector-python
python-dotenv
pandas
scikit-learn
numpy
matplotlib